Split Model type class from vector into separate module
Rename old Model module into MatModel
