+++ vecvec-lapack/tests/TST/LinSolve.hs
@@ -42,7 +42,7 @@ import Vecvec.LAPACK.Matrix.Dense          (Matrix)
 import Vecvec.LAPACK.Matrix.Dense          qualified as Mat
 import Vecvec.LAPACK.LinAlg
 
-import TST.Tools.Model
+import TST.Tools.MatModel
 import TST.Tools.Util
